Major commit
Sorry been working on the listing pages. The design is there, working fully cross browser/device. Spend the last few hours working on a Utils/Modal.

Im happy to spend time on the Modal window to make it as robust as possible, for use throughout the site. Current version only checked in webkit, also needs work on removing from dom after our beautiful user clicks it away. Not to keen on the animation either. aaaah